+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/RequestParameters.kt
@@ -1,6 +1,14 @@
 package at.asitplus.openid
 
-interface RequestParameters
+interface RequestParameters {
+    val responseType: String?
+    val nonce: String?
+    val clientId: String?
+    val redirectUrl: String?
+    val audience: String?
+    val state: String?
+    val transactionData: Set<String>?
+}