+++ rqes-data-classes/src/commonMain/kotlin/at/asitplus/rqes/serializers/RequestParameterSerializer.kt
@@ -3,7 +3,6 @@ package at.asitplus.rqes.serializers
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.CscAuthenticationRequestParameters
 import at.asitplus.openid.RequestParameters
-import at.asitplus.rqes.Hashes
 import at.asitplus.rqes.SignatureRequestParameters
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.json.JsonContentPolymorphicSerializer